+++ package.json
@@ -10,31 +10,31 @@
   },
   "scripts": {
     "build": "npm run clean && npm run build:data && webpack --progress --colors --bail",
-    "build:data": "babel-node scripts/build-data",
+    "build:data": "node scripts/build-data.mjs",
     "clean": "rimraf ./dist ./locales && mkdirp dist locales",
     "lint": "npm run lint:js && npm run lint:json",
-    "lint:js": "eslint . --ext .js",
+    "lint:js": "eslint . --ext .js,.mjs,.cjs",
     "lint:json": "jshint -e .json www editor/blocks editor/extensions editor/interface editor/paint-editor",
     "prepare": "husky install",
-    "pull:blocks": "babel-node scripts/tx-pull-editor scratch-editor blocks ./editor/blocks/",
+    "pull:blocks": "node scripts/tx-pull-editor.mjs scratch-editor blocks ./editor/blocks/",
     "pull:editor": "npm run pull:blocks && npm run pull:extensions && npm run pull:paint && npm run pull:interface",
-    "pull:extensions": "babel-node scripts/tx-pull-editor scratch-editor extensions ./editor/extensions/",
+    "pull:extensions": "node scripts/tx-pull-editor.mjs scratch-editor extensions ./editor/extensions/",
     "pull:help": "npm run pull:help:names && npm run pull:help:articles",
     "pull:help:articles": "./scripts/tx-pull-help-articles.js",
     "pull:help:names": "./scripts/tx-pull-help-names.js",
-    "pull:interface": "babel-node scripts/tx-pull-editor scratch-editor interface ./editor/interface/",
-    "pull:paint": "babel-node scripts/tx-pull-editor scratch-editor paint-editor ./editor/paint-editor/",
-    "pull:www": "babel-node scripts/tx-pull-www ./www",
-    "push:help": "./scripts/tx-push-help.js",
+    "pull:interface": "node scripts/tx-pull-editor.mjs scratch-editor interface ./editor/interface/",
+    "pull:paint": "node scripts/tx-pull-editor.mjs scratch-editor paint-editor ./editor/paint-editor/",
+    "pull:www": "node scripts/tx-pull-www.mjs ./www",
+    "push:help": "./scripts/tx-push-help.mjs",
     "sync:help": "npm run push:help && npm run pull:help",
     "test": "npm run lint:js && npm run validate:editor && npm run validate:www && npm run build && npm run lint:json",
     "update": "scripts/update-translations.sh",
-    "validate:blocks": "babel-node scripts/validate-translations ./editor/blocks/",
+    "validate:blocks": "node scripts/validate-translations.mjs ./editor/blocks/",
     "validate:editor": "npm run validate:blocks && npm run validate:extensions && npm run validate:interface && npm run validate:paint",
-    "validate:extensions": "babel-node scripts/validate-translations ./editor/extensions/ && babel-node scripts/validate-extension-inputs",
-    "validate:interface": "babel-node scripts/validate-translations ./editor/interface/",
-    "validate:paint": "babel-node scripts/validate-translations ./editor/paint-editor/",
-    "validate:www": "babel-node scripts/validate-www ./www"
+    "validate:extensions": "node scripts/validate-translations.mjs ./editor/extensions/ && node scripts/validate-extension-inputs.mjs",
+    "validate:interface": "node scripts/validate-translations.mjs ./editor/interface/",
+    "validate:paint": "node scripts/validate-translations.mjs ./editor/paint-editor/",
+    "validate:www": "node scripts/validate-www.mjs ./www"
   },
   "repository": {
     "type": "git",
