+++ src/index.mjs
@@ -1,0 +1,3 @@
+import localeData from './locale-data.mjs';
+import locales, {localeMap, isRtl} from './supported-locales.mjs';
+export {locales as default, localeData, localeMap, isRtl};
