+++ webpack.config.js
@@ -22,8 +22,8 @@ module.exports = {
     },
     entry: {
         l10n: './src/index.js',
-        supportedLocales: './src/supported-locales.js',
-        localeData: './src/locale-data.js'
+        supportedLocales: './src/supported-locales.mjs',
+        localeData: './src/locale-data.mjs'
     },
     output: {
         path: path.resolve(__dirname, 'dist'),
